requirements.txt
Flask-Script
Flask-SQLAlchemy
alembic
